feat(DTFS2-7222): review markups
