--- trade-finance-manager-api/src/v1/__mocks__/mock-amendment.js
@@ -1,12 +1,0 @@
-const { MOCK_TASKS } = require('./mock-amendment-tasks-assign-by-team');
-
-const MOCK_AMENDMENT = {
-  amendmentId: '66310991135a69c4c2af734a',
-  facilityId: '65af9a9237544e9808723523',
-  dealId: '65af9a9237544e9808723520',
-  tasks: MOCK_TASKS,
-};
-
-module.exports = {
-  MOCK_AMENDMENT,
-};
