+++ trade-finance-manager-api/src/v1/__mocks__/mock-facility-snapshot.ts
@@ -1,4 +1,4 @@
-import { Facility, FACILITY_TYPE, TfmFacilityAmendment } from '@ukef/dtfs2-common';
+import { Facility, FACILITY_TYPE } from '@ukef/dtfs2-common';
 import { ObjectId } from 'mongodb';
 
 export const MOCK_FACILITY_SNAPSHOT: Facility = {
@@ -34,27 +34,4 @@ export const MOCK_FACILITY_SNAPSHOT: Facility = {
   hasBeenIssuedAndAcknowledged: null,
   canResubmitIssuedFacilities: null,
   unissuedToIssuedByMaker: {},
-};
-
-export const MOCK_AMENDMENT: TfmFacilityAmendment = {
-  version: 1,
-  amendmentId: new ObjectId(),
-  facilityId: new ObjectId(),
-  dealId: new ObjectId(),
-  createdAt: 1723653619,
-  updatedAt: 1723653634,
-  status: 'Completed',
-  tfm: {
-    amendmentExposurePeriodInMonths: null,
-    coverEndDate: 1794418807,
-    exposure: {
-      exposure: '0.50',
-      timestamp: 1828546829000,
-      ukefExposureValue: 0.5,
-    },
-    value: {
-      currency: 'GBP',
-      value: 5,
-    },
-  },
 };
