+++ trade-finance-manager-api/src/v1/rest-mappings/mappings/gef-facilities/mapGefFacilityDates.test.js
@@ -132,6 +132,9 @@ describe('mapGefFacilityDates', () => {
       const facilityWithAmendments = {
         facilitySnapshot: {
           ...mockFacility.facilitySnapshot,
+          isUsingFacilityEndDate: undefined,
+          bankReviewDate: undefined,
+          facilityEndDate: undefined,
         },
         amendments: [
           {
