+++ trade-finance-manager-ui/server/routes/case/index.test.js
@@ -1,13 +1,11 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { expect, jest } from '@jest/globals';
-import { TEAM_IDS } from '@ukef/dtfs2-common';
 import { get, post } from '../../test-mocks/router-mock';
 import caseController from '../../controllers/case';
 import partiesController from '../../controllers/case/parties';
 import underwritingController from '../../controllers/case/underwriting';
 import activityController from '../../controllers/case/activity';
 import amendmentsController from '../../controllers/case/amendments';
-import { validateUserTeam } from '../../middleware';
 
 describe('routes - case', () => {
   afterEach(() => {
